Remove includes as it's buggy
@heathd and I have dug into this more - when using `includes(:policy_topics)`
the `class_name` directive for the association isn't used, which causes this to
load WorldwidePriority objects instead of the Policy objects we actually
wanted.

Removing this will make the query slower but won't hit this bug, which
is arguably a bug in ActiveRecord.

The Edition which triggers this the first instance of this error is 160804.