Remove redundant checks for duplicate errors
These tests are now redundant as the `#errors` attribute of an object is
not modified by `#valid_as_draft?`. To access these errors you can use
`#errors_as_draft` (as is done in the only place in the code where it is
needed).

The tests became redundant in b1828089bc9a43ec57c420cd22129647d9cf43e2
because the method of checking for errors when converting to draft was
changed (part of the rails 3.2 migration).