Stop organisations in meta footer clashing IDs
Because we use content_tag_for, the IDs are automatically generated. We
can work around this using the prefix argument, but that should only be
overridden in the template.

The problem this causes is that there are two things with IDs of
"organisation_1234" on the publications page, as the organisation is
shown more than once.