Cleanup blank consultation outcomes
In today's view of the world, we want consultation outcomes and public
feedback to always have some summary text. But because we have legacy
responses that might not have a summary, we want to losen the
validation a bit so that they remain valid. To this end, we only
validate the summary is present if there are no attachments present.

https://www.pivotaltracker.com/story/show/54897632