DRY up use of scopes to find latest editions
There are plenty of controllers that want to find the latest few
editions of a specific type (e.g. the 3 latest policies that have been
published in a world location). There's plenty of duplication in those
chained scopes, so we can tidy things up a fair bit by moving them into
a method and setting appropriate defaults.