Rework association between series and documents
Instead of linking series with editions, we now link directly to the
document. This removes the need to copy across the association with
each new edition and greatly simplifies the code in general. We no
longerr list series documents grouped by type on frontend - this added
little value to the user.