Redesign admin edition show page
This is the first pass at redesigning the edition show page to
give it more structure. The main change is you can no longer
preview documents or document metadata from the admin – you
have to go through to the frontend preview page. This is
because the admin preview is misleading. You aren't actually
previewing how it will look.

The other major change is to how translations are done. They
now are listed in a table instead of displaying the full
preview in the admin, and a link to preview them on frontend
has been added.

Most of the tests that have been removed are ones that test
for metadata that is no longer there.

Some other notes

- Edition#previewable? is now just `not published?`.
  There was an odd concept of previewable before.
- I have started to clean up some of the admin CSS to match how
  it is done on the frontend. Loads more to do.

https://www.pivotaltracker.com/story/show/53973095