Improve document series adding feedback
Search results that are already part of the series will be identified
as such. We also update the search results to reflect once a document
has been added. To do this, I've moved the search action out of the
main editions controller. I had hoped to make this a generic search
action, but the mustache template was already tied to document series
management, so it was only a half measure anyway.