Use public_timestamp for consultation email notifications
For a new consultation the public_timestamp will be set to the
opening_on date of the consultation. I think that when this notifier
code was written, that was not the case, which is why there was a special
exception for consultations in the code. However, it's not needed any
more.

This means that when adding a consultation to gov.uk, if the opening_on
date is in the past then no notification will go out.

For a major change to a consultation, the public_timestamp will be set
to the major change date, and so all major changes will cause an email
notification and cause the document to appear at the top of
chronological lists. This is the desired behaviour.

https://www.pivotaltracker.com/story/show/55288280