Fix html slug test
The test was coupled to the title of the html set by the factory. This
patch breaks that dependency by setting the title explicitly.
This test was broken as the title in the (now removed) trait was
different to the title in the html_version factory that is used instead.