Tidy up after associating topics directly with editions
Editions used to be associated with topics via policies. When we rolled
out the code that associated editions with topics we ran a rake task to
copy the old associations over. Before the rake task finished, the code
that is being removed in this commit enabled the site to use either
association to find out which topics an edition was associated with.

This reverts commit ec8a4b15360571b1db36d721238e7b5b72626cbd.

Story: https://www.pivotaltracker.com/story/show/55631686