Add models to track grouping of document series
This commit also contains a migration to copy data from the
document_series_memberships table into the new tables, and a pending
data migration to drop the document_series_memberships table once the
code is running successfully in production.

When the groups are retrieved from a series, they're returned in the
order in which they were added. Similarly for the documents in a group.