Delete groups from doc series
If a user tries to delete a group with documents in it, they're told
that they need to get rid of the documents first. The message appers on
a page rendered by the `#delete` action, which also doubles up as the
confirmation "dialog" if the group happens to be empty.

There's no server side checking to ensure that only empty groups are
deleted; it seems like it'd be unnecessary complication given the access
that a user who can delete an empty group already has to the contents of
the group.