Kill off PublicationType::Consultation
I see no reason for this type to exist anymore - no Publication should
have this type, and Consultations are identifiable without it. This
required some slight tweaking to the MySQL front-end filterer.

(NOTE: the MySQL filterer isn't smart enough to distinguish between
open and closed consultations. That was the case before these changes)