Index document_series correctly in tests
Indexing occurs at publication time. If you create a
published_publication using the factory then it the indexing trigger
would fire at that moment, before it would be associated with the
document_series.

I think this worked previously because the controller was loading the
object instances from the database by ID, so it was getting the
persisted object rather than the value actually stored in the index.

To ensure that the document_series value is stored in the index, we need
to publish after setting up the whole object graph.

Also, add document_series to not_quite_as_fake_search

We need to tell it to handle this field. This is a bit non-DRY but hard
to sync up with the canonical definition (in rummager mappings).