Handle nil operational_field
A FatalityNotice requires operational_field to be set, so it's
reasonable to assume that a FatalityNotice would have a value for
operational_field if the method is defined.

However, the Result object will define this method for all edition
types, even if they don't have operational_field. For those other
edition types the value will be empty. Therefore so we need to check
that it has a value as well as the method being defined.