Fix schema.rb to reflect schema in production
Because code is merged and deployed out of sequence, migrations may
get run out of order on preview/production. If a developer then runs
migrations locally in order, we end up with a discrepency between the
local schema.rb and our production schema. This commit puts schema.rb
back to the same state as our production database. If you're local
database keeps moving this column in schema.rb, then the remedy is to
pull down the latest dump from preview and not to commit this change.
Your local database will then be in sync with production and schema.rb