Don't compile mustache when running cucumber
We needed to do this while there was cucumber coverage of code that uses
mustache, to ensure that the latest templates were available to the
cucumber scenarios.

I've just modified the admin UI that used cucumber (the doc code that
finds docs when adding them to a doc series), so we no longer need this
step. I think we're better off without it as (should cucumber not
finish) compiled templates will be left on disk, and they may
accidentally get committed.