Use Rails helper method to generate grouped select
Perhaps this was built manually as the `with_translations_for` method was
involved? It doesn't appear to have mattered, as including translations
in the AREL code seems to have done the job nicely; the HTML produced is
identical, other than for the order of the optgroup tags (which come out
in alphabetical order with this new approach).