Remove Classification#recently_changed_documents
The previous implementation loaded into memory all documents relating to
the classification. We would then order by timestamp and take the first
three. Clearly extremely inefficient.

Since Graham's changes in d9c01f02 editions are directly associated with
topics as well as being associated with policies. Therefore we no
longer need to traverse the policy association to load the editions.

One nuance is that loading via topics directly will pull in other
edition types (in particular WorldLocationNewsArticles) which previously
would not have shown up on the Topic 'latest' feed. These are not
sufficiently relevant to a general audience and so they are explicitly
excluded from the 'latest' list on topics.

Deleted quite a few no-longer relevant tests as a result of this.