Display custom logo on org, about and corporate info pages
For organisations with a custom logo we drop the coloured border, the
crest, and the name of the organisation from the <h1> tag. We replace
the logo with an image that GDS will have edited to a suitable size, to
work well on the site. The intent is that the ability to upload images
will be limited to GDS staff (possibly via process, possibly via access
controls that we might add at a future date).

When displaying the name of an org in a list we often include a small
crest and a coloured border. For organisations with a custom logo we
can no longer do this, so we drop the border and don't set a background
image.