remove diffy and use google's js diff library
This enables editors to see changes inline more akin to MS words
'redline'. The unix diff library duplicates lines which can make it
harder to see the actual change. To achieve this we use the default
settings of dpm which are:

  Diff_EditCost = 4;
  Match_Threshold = 0.5;

And the diff_cleanupEfficiency algorithm.

The API docs are here:
https://code.google.com/p/google-diff-match-patch/