Add separate action for re-ordering features
The way it was done previously using the classifications#update action
meant that any related classifications were being cleared every time
the features were-reordered. This mirrors what we already do for
re-ordering attachments.

I also updated the factory for a classification featuring to ensure
the featured edition is published, otherwise the association on
classification does not return them as it has an explicit conditiion
set to only load those with published editions