Remove dodgy rescue_from
This is a really nasty bit of code that can and will cause all sorts
of potential problems. We should never simply swallow
ActiveRecord::RecordInvalid if can help it as it doesn't guarantee
the data will be in a sane state. It also makes issues incredibly hard
to debug. For example, recently an edition was being published. The
edition in question was valid and successfully published, but becuase
previous editions were not valid, the `archive_previous_editions`
callback failed, raising an exception. But instead of that bubbling to
the top and everything being rolled back with a 500 error, the
exception was caught by the rescue_from, the edition was published,
and a cryptic error message was shown to the user saying the edition
could not be published, even though it was now in a published state.

Admittedly it's not ideal that users will potentially see 500 errors,
but we should avoid that by aggressivly catching all the instances
where we are trying to perform invalid transitions, rather than by
brushing things under the carpet in this way.