Rely on Whitehall::Authority for access limiting
Step one of removing all authorisation-related code from the Edition
model. Thanks to @hlame's steller work, all the logic for access
limiting editions already exists in the enforcement rules within
Whitehall::Autthority and are also thoroughly tested, so this was a
relatively straightforward process. The access limiting method on
Attachment was also factored out for completeness.