Replace associations with scopes and tidy up
Too many associations where leveraging scopes will do - this gets of
half a dozen.

I've also renamed the unpublished_edition method to
non_published_editions as we the concept on an unpublished edition is
now a thing in the app and they're not quite the same thing.

Finally, this gets rid of some tests around topics and policies that
are no longer relevant as topics are now directly assigned to
editions, rather than inferred through policies.