Decouple subscribers from publishing services
The edition service objects should not need to have knowledge of the
callbacks/subscribers that are interested in their outcome. Instead,
we rely on another object to coordinate the subscription and
notification of interested parties - The Edition::ServiceCoordinator.
This uses the simple queue pub/sub class from within ActiveSupport to
push events back and forth.