Simplify resetting version number after unpublish
The decrement_version_number method is used when unpublishing a
document to put the version number back to it's pre-published state.
So what we actually want to do is reset it, rather than decrement it.
This is as simple as setting it the version number of the previous
edition, or nil if one doesn't exist. I've repurposed the
edition_to_diff_against method to retrieve the previous edition.