Separate service object for archiving editions
It felt wrong that there was a single service object for archiving
and unpublishing so I've split them out (note that archived editions
will remain in the search index for now, as the search status of
archived editions is being dealt with in a separate story). In an
ideal world, we'd also have a separate controller action for each, but
because of the way the UI works, this would be quite tricky, so
leaving it as a single controller action for now.