Handle old admin supporting page paths in govspeak
Supporting pages are now editions so have admin paths of the form
/admin/supporting-pages/:id, but some existing documents will contain
links of the form /admin/editions/:policy_id/supporting-pages/:slug.

Note: supporting pages which have had their slug changed as part of the
supporting page migration will be incorrectly matched by this so the
links will be wrong.