Get rid of legacy edition featuring code
Editions featured on organisations are now done via Features/
FeatureLists. This blows away the code and data that we no longer
need.

Note: The actual files will still be on the filesystem of the servers
at uploads/system/uploads/edition_organisation_image_data. Whether
they should be deleted or not (they may still be linked to online) is
for someone else to decide.