Expand AttachmentVisibilty functionality
This is so the appropriate records can be loaded to preview a CSV
attachment.

I've added some docs explaining the utility of this class as it's a
strange thing that only exists because of existing debt around the way
we route attachments. Hopefully it will act as a handy signpost to
future-developer and give them some context as to the reasons why it
exists.