Ensure Attachment upload fields are present
For various reasons, there are Attachment instances that are neither
a FileAttachment or an HtmlAttachment *. This ensures we always show
the file upload fields for non-HtmlAttachments.

* Firstly, the importer creates Attachments instead of
FileAttachments. This is easily rectified, but the main issue is that
STI and mass assignment don't play nicely, as the "type" attribute is
protected from mass-assignment in Rails < 4. This means that whenever
attachments are manipulated using mass-assignment (i.e. using the old
attachments interface on PolicyAdvisoryGroups and
CorporateInformationPages and), the type will get lost.