Add note on abstract-ness of Attachment class
I tried to add code that prevents these from ever being saved, but
unfortunately that isn't very straighforward -The app currently uses
mass-assignment to enable the manipulation of attachments on
PolicyAdvisoryGroups and CorporateInformationPages, and
mass-assignment in Rails < 4 will not preserve the STI type attribute.

Either the admin interface for managing attachments on these models
should be updated to match the new one used on editions, or we should
upgrade to Rails 4, or ideally, both.