Fix comparison of AuditEntry objects
The <=> method incorrectly used assignment when trying to check if the
sort_priority of the two objects was equal.

I've updated it to use a simpler approach for multi-field sorting by
putting the comparison fields into arrays and comparing those. I've also
removed the object comparison as the <=> method inherited from
ActiveRecord::Base returns nil if the records aren't the same class
breaking the sort.