Speculative fix for intermittently failing feature
Scenario in question is features/document-collections.feature:33

This scenario often fails on the step:
"And I see that "Feeding Wombats" is before "Wombats of Wimbledon" in the document collection"

I'm thinking that this might be because the post made in
"And I move "Feeding Wombats" before "Wombats of Wimbledon" in the document collection"
was aborted before completing by navigating straight to the preview page. If I'm right,
waiting for the loading spinner to go should fix this.