Second step: repair version numbers/change notes
This fixes the major/minor flag, change notes and numbering of
supporting pages. The intention is the following:

1. The first migrated version is the only one that is designated as
a major change.
2. Change notes for migrated versions are cleared
3. All migrated versions get a published_version of "1.0"

The reasoning is as follows:

We want to be able to include the change history of supporting pages
along side the parent policy, but at the moment, all editions are
marked as being  a major change and have adopted the change notes of
the policy. This means that a combined change history would include
duplicate change notes and would include each revision of the
supporting pages, regardless of whether there were any actual changes
to the page. The above chnages mean that only the very first migrated
edition will be included in the change history (the only major
change). This will result in a sane history, as we can rely on the
parent policy supplying any relevant change notes for the supporting
pages created prior to migrating them to editions.

The change notes and minor_change flag are not updated for any
supporting page editions created after the migration, as these would
have been revised by an editor and should therefor have sensible
minor_change/change note values. We do however re-number these
editions, as they will be incorrectly numbered - because the migrated
editions had nil version numbers, the revised editions would have been
designated as the first published version and given a version number
of "1.0".