Exclude already-covered supporting page chganges
Those supporting pages that were migrated will share a public
timestamp with their parent policy, which should already have an
appropriate change note. Therefore, we exclude any supporting change
changes from the history that share a public timestamp with a change
in the main policy document.

This also adjusts the wording of the change history generated for the
first published supporting page.