Remove local government document filter
The filter hasn't been used on production for a long while, but we were
still keeping the search index up to date for it. This meant reindexing
every edition related to a policy when that policy was reindexed.

This commit pulls out the filter code and also the indexing code that
was used to enable it.