Show locale-appropriate content on activity feeds
Updates the activity on policies and worldwide priorities to only list
content that has been translated into the current locale. Does the
same for the policy/announcement feeds on people pages as well.

Note that as well as un-stubbing the PeopleControllerTests,  I've also
converted the tests for RolePresenter and PersonPresenter so that they
no longer inherit from PresenterTest; in both cases, database
connections were being disabled and everything stubbed. Great in
theory for faster tests, but in reality it just makes testing our
tangled domain model really (really) hard.