Handle timeouts and 502s from GDS API
We receive dozens of exception notifications every day because
rummager search requests are timing out. This will prevent the
notifications being sent, whilst ensuring a 500 response is still
returned so that cached results can be served in production.

This is definitely masking the symptom rather than fixing the cause,
but as things stand, legitimate exceptions are getting lost in the
noise.