Rename existing DocumentMappings
The existing export is inconsistent and odd. It includes tens of thousands of
rows that serve no purpose, has inconsistent behaviour for different data types
and includes duplicate rows where there are multiple editions of a given item.

Changing it is therefore both risky and undesirable because it may change the
behaviour of the redirector 'munge' process, which although we are moving away
from, we need for the time being.