Clean up response factory
We shouldn't be using factories to build abstract responses, just the
concrete subclasses (consultation outcome and consultation public
feedback). If an abstract response really is wanted it can be built
using Response.new.

Also responses don't make sense without a consultation, so make sure we
build one in the factory.