Migration to fix duplicate files on editions
Any published editions with attachments that have duplicate filenames
are re-editioned and updated to make their attachment files unique
before being published as a minor change. This uses a modified
EditionForcePublisher that skip validations and resets the force
published flag. This is because there are a number of old editions
which do not have any topics set. It also means we do not need to
worry about waiting for the virus scanner to finish scanning before
publishing.