Remove #completed_fact_check_requests.markLinkedAnchor
The result of this call is to find an element by id matching the
hash portion of the current location (??!) and to add a 'linked' class 
to it. When this section is visible, the hash portion is #fact_checking,
and no element in the source code matches that. I'm assuming this is
legacy and removing it.