Tidy edition perm checks in attachments controller
Ideally the permissions system would be able to handle objects other
than editions so that can?(:see, attachable) would work for any type of
attachable. Given that it doesn't, and fixing the permissions system is
a bigger job, this tidies up the before filters so we only have a single
one related to edition checks instead of four.