Stop HtmlAttachment slug generation complaining
friendly_id doesn't like generating slugs for HtmlAttachment because
the scope it gets given is on the base Attahment class, and that hasn't
had the necessary friendly_id helpers mixed into it. By forcing the
scope to one on the HtmlAttachment class itself, we keep friendly_id's
slug generator happy and also tighten up the slug scopping so that
html attachments' slugs are scoped to the owning document.