Remove #readonly override from join models
Rails 4 changes the behaviour of ActiveRecord such that a model that
is marked as "readonly" can no longer be destroyed (see
https://github.com/iGEL/rails/commit/a8b370f5c2fed87001e363d18ce60a47c1748885).
This prevents EditionRelations and ClassificationRelations from being
destroyed. Looking back through the git history, it looks like this
was originally put in as a way to prevent these join models from being
modified once they were created. I can't think of any good reason why
we should be enforcing this at the model level so I'm simply removing
the readonly designation from the model.